+++ divio_cli/validators/addon.py
@@ -1,4 +1,5 @@
-import imp
+import importlib.machinery
+import importlib.util
 import os
 import shutil
 import time
@@ -15,6 +16,20 @@ from .common import load_config, validate_package_config
 ADDON_REQUIRED_CONFIG_KEYS = ("package-name",)
 
 
+def load_source(modname, filename):
+    """Compatibility function to replace usage of `imp` module."""
+    loader = importlib.machinery.SourceFileLoader(modname, filename)
+    spec = importlib.util.spec_from_file_location(
+        modname, filename, loader=loader
+    )
+    module = importlib.util.module_from_spec(spec)
+    # The module is always executed and not cached in sys.modules.
+    # Uncomment the following line to cache the module.
+    # sys.modules[module.__name__] = module
+    loader.exec_module(module)
+    return module
+
+
 def validate_aldryn_config_py(path):
     aldryn_config_path = os.path.join(path, "aldryn_config.py")
     if os.path.exists(aldryn_config_path):
@@ -29,7 +44,7 @@ def validate_aldryn_config_py(path):
 
                     # randomizing source name
                     source = f"aldryn_config.config_{int(time.time())}"
-                    module = imp.load_source(source, temp_path)
+                    module = load_source(source, temp_path)
 
                 # checking basic functionality of the Form
                 form = module.Form({})