+++ CHANGELOG.rst
@@ -20,6 +20,7 @@ x.x.x (x-x-x)
 * Removed template release commands confirm prompt in non-interactive mode.
 * Migrated from `setup.py`` to `pyproject.toml` for packaging.
 * Added support for username/password authentication on external repositories during app creation.
+* Removed usage of `imp` module to better support python 3.12.
 
 3.17.2 (2023-03-16)
 -------------------